Reduce redundancy in mamba2 blocks
